fix(seguridad): el guard de ventana community tambien en long
Ultimo fleco del arreglo anterior, senalado por la auditoria adversaria: el
guard que acota la peticion a la ventana de comunitarias seguia calculando
reqLast en int. Con un offsetBase enorme la suma se desborda a negativo y las
dos comparaciones de la ventana lo dan por bueno, asi que era la unica
aritmetica del handler que seguia dependiendo de que UnlockChainWire hubiera
acotado antes. En long el guard vale por si solo, que es justo lo que decia
pretender el comentario del bucle de abajo.

Sin cambio funcional para el trafico legitimo: los offsets community reales van
de 21 a 36 en el peor caso (mesa de 10 con run-it-twice), muy lejos de
cualquier borde.

RESIDUAL CONOCIDO, no arreglado aqui: al cortar el offset fuera de rango en el
parser, ese REQ se rechaza como malformacion estructural (WARNING silencioso)
en vez de disparar el lockdown que si dispara un offset entre 52 y 4095. La
peticion se rechaza igual en ambos casos, pero un host tramposo que pruebe por
encima del tope no deja rastro visible. Queda anotado en el informe.

+++ src/main/java/com/tonikelope/coronapoker/WaitingRoomFrame.java
@@ -2815,7 +2815,13 @@ public class WaitingRoomFrame extends JFrame {
                                                                             return;
                                                                         }
                                                                         if (commRange != null) {
-                                                                            int reqLast = it.offsetBase + it.chains.size() - 1;
+                                                                            // long por el mismo motivo que el bucle de abajo: offsetBase
+                                                                    // viene del wire y en int la suma se desborda a negativo,
+                                                                    // con lo que este guard de ventana daria por bueno un
+                                                                    // offset enorme. UnlockChainWire ya lo acota al parsear;
+                                                                    // esta era la ultima aritmetica del handler que seguia
+                                                                    // dependiendo de aquello.
+                                                                    long reqLast = (long) it.offsetBase + it.chains.size() - 1;
                                                                             if (it.chains.isEmpty() || it.offsetBase < commRange[0] || reqLast >= commRange[0] + commRange[1]) {
                                                                                 LOGGER.log(Level.SEVERE,
                                                                                         "ZERO-TRUST: REQ_SRA_UNLOCK_CHAIN offset {0}(+{1}) outside phase {2} community slots [{3},{4}) — host reading the future board, refusing",
